layer manager, reorganize files and folders, upgrade packages, remove zod
Sorry for the giant commit with unrelated things.

Removed zod because it inferred a different type from a Record of an enum than typescript normally would, causing over an hour of type debugging that would not have happened otherwise. And it was kind of a premature optimization to begin with. If I need to validate locally stored data in the future before merging into the store, I could do it manually, or reintroduce zod, or something else.